requirements.txt
blessed==1.20.0
defusedxml==0.7.1
duckdb==0.7.1
fonttools==4.39.3
fpdf2==2.7.3
inquirer==3.1.3
numpy==1.24.3
pandas==2.0.0
Pillow==9.5.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2023.3
readchar==4.0.5
six==1.16.0
tzdata==2023.3
wcwidth==0.2.6
